# bpo-45636: Merge `BINARY_*`/`INPLACE_*` into `BINARY_OP`/`INPLACE_OP`