--- a/Include/cpython/object.h
+++ b/Include/cpython/object.h
@@ -123,7 +123,7 @@ typedef struct {
     binaryfunc nb_xor;
     binaryfunc nb_or;
     unaryfunc nb_int;
-    void *nb_reserved;  /* the slot formerly known as nb_long */
+    unaryfunc nb_reserved;  /* the slot formerly known as nb_long */
     unaryfunc nb_float;
 
     binaryfunc nb_inplace_add;