--- a/Include/internal/pycore_abstract.h
+++ b/Include/internal/pycore_abstract.h
@@ -16,6 +16,9 @@ _PyIndex_Check(PyObject *obj)
     return (tp_as_number != NULL && tp_as_number->nb_index != NULL);
 }
 
+PyObject *_PyNumber_Op(PyObject *o1, PyObject *o2, unsigned op);
+PyObject *_PyNumber_InPlaceOp(PyObject *o1, PyObject *o2, unsigned op);
+
 #ifdef __cplusplus
 }
 #endif