--- a/Lib/dis.py
+++ b/Lib/dis.py
@@ -7,6 +7,7 @@ import io
 
 from opcode import *
 from opcode import __all__ as _opcodes_all
+from opcode import _nb_ops
 
 __all__ = ["code_info", "dis", "disassemble", "distb", "disco",
            "findlinestarts", "findlabels", "show_code",
@@ -27,6 +28,14 @@ MAKE_FUNCTION = opmap['MAKE_FUNCTION']
 MAKE_FUNCTION_FLAGS = ('defaults', 'kwdefaults', 'annotations', 'closure')
 
 LOAD_CONST = opmap['LOAD_CONST']
+
+BINARY_OP = opmap['BINARY_OP']
+BINARY_OPS = {i: name for i, (_, name) in enumerate(_nb_ops)}
+
+INPLACE_OP = opmap['INPLACE_OP']
+INPLACE_OPS = {i: f"{name}=" for i, (_, name) in enumerate(_nb_ops)}
+
+del _nb_ops
 
 def _try_compile(source, name):
     """Attempts to compile the given source, first as an expression and
@@ -446,6 +455,10 @@ def _get_instructions_bytes(code, varname_from_oparg=None,
             elif op == MAKE_FUNCTION:
                 argrepr = ', '.join(s for i, s in enumerate(MAKE_FUNCTION_FLAGS)
                                     if arg & (1<<i))
+            elif op == BINARY_OP:
+                argrepr = BINARY_OPS[arg]
+            elif op == INPLACE_OP:
+                argrepr = INPLACE_OPS[arg]
         yield Instruction(opname[op], op,
                           arg, argval, argrepr,
                           offset, starts_line, is_jump_target, positions)