--- a/Lib/opcode.py
+++ b/Lib/opcode.py
@@ -205,16 +205,43 @@ def_op('CALL_METHOD_KW', 166)
 
 del def_op, name_op, jrel_op, jabs_op
 
-_nb_ops = [
-    ("NB_AND", "&"),
-    ("NB_FLOOR_DIVIDE", "//"),
-    ("NB_LSHIFT", "<<"),
-    ("NB_MATRIX_MULTIPLY", "@"),
-    ("NB_OR", "|"),
-    ("NB_RSHIFT", ">>"),
-    ("NB_SUBTRACT", "-"),
-    ("NB_TRUE_DIVIDE", "/"),
-    ("NB_XOR", "^"),
+_nb_slots = [
+    # nb_add
+    (1, "nb_subtract", "-"),
+    # nb_multiply
+    # nb_remainder
+    # nb_divmod
+    # nb_power
+    # nb_negative
+    # nb_positive
+    # nb_absolute
+    # nb_bool
+    # nb_invert
+    (11, "nb_lshift", "<<"),
+    (12, "nb_rshift", ">>"),
+    (13, "nb_and", "&"),
+    (14, "nb_xor", "^"),
+    (15, "nb_or", "|"),
+    # nb_int
+    # nb_reserved
+    # nb_float
+    # nb_inplace_add
+    (20, "nb_inplace_subtract", "-="),
+    # nb_inplace_multiply
+    # nb_inplace_remainder
+    # nb_inplace_power
+    (24, "nb_inplace_lshift", "<<="),
+    (25, "nb_inplace_rshift", ">>="),
+    (26, "nb_inplace_and", "&="),
+    (27, "nb_inplace_xor", "^="),
+    (28, "nb_inplace_or", "|="),
+    (29, "nb_floor_divide", "//"),
+    (30, "nb_true_divide", "/"),
+    (31, "nb_inplace_floor_divide", "//="),
+    (32, "nb_inplace_true_divide", "/="),
+    # nb_index
+    (34, "nb_matrix_multiply", "@"),
+    (35, "nb_inplace_matrix_multiply", "@="),
 ]
 
 _specialized_instructions = [