--- a/Misc/NEWS.d/next/Core and Builtins/2021-10-27-15-14-31.bpo-45636.K2X7QS.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2021-10-27-15-14-31.bpo-45636.K2X7QS.rst
@@ -1,0 +1,2 @@
+Merge most ``BINARY_*`` and ``INPLACE_*`` instructions into unified
+:opcode:`BINARY_OP` and :opcode:`INPLACE_OP` implementations.