… Doc/c-api/typeobj.rst
… Doc/library/dis.rst
… Doc/whatsnew/3.11.rst
… Lib/test/test_compile.py
… Lib/test/test_dis.py
… Misc/NEWS.d/next/C API/2021-11-01-08-56-18.bpo-45636.c8DIYB.rst
… Python/opcode_targets.h
… Tools/scripts/generate_opcode_h.py