<!-- issue-number: [bpo-45636](https://bugs.python.org/issue45636) -->
https://bugs.python.org/issue45636
<!-- /issue-number -->