@brandtbucher 2021-10-29
Thinking about this a bit more... I'd like to merge this as-is, then open a follow-up PR with the more complex implementation that performs fewer memory reads. I'd like to at least see some movement on the benchmarks before going that route.