@markshannon 2021-11-02
I can confirm the [slowdown](https://gist.github.com/markshannon/0a6723e21ac1a2d44035388b2fea0149), which is a bit puzzling.
If your original design is faster, then maybe we should go with that. It is a bit cleaner.
Ultimately I expect specialization to make the performance of the base version unimportant.