@brandtbucher 2021-11-05
Closing in favor of #29418, which is based on an earlier commit on this branch (and has no perf impact).